% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca_retrospective_v5.0_proteome_gct}
\alias{brca_retrospective_v5.0_proteome_gct}
\title{Proteome example GCT}
\format{
\subsection{'brca_retrospective_v5.0_phosphoproteome_gct'}{

Formal class 'GCT' from package \code{cmapR} with 7 slots:
\describe{
\item{mat}{data matrix}
\item{rid}{row ids}
\item{cid}{column ids}
\item{rdesc}{row description dataframe}
\item{cdesc}{column description dataframe}
\item{version}{GCT version}
\item{src}{GCT source file}
}
}
}
\source{
https://github.com/broadinstitute/PANOPLY/blob/dev/tutorial/tutorial-brca-input.zip
}
\usage{
data(brca_retrospective_v5.0_proteome_gct)
}
\description{
A GCT containing the BRCA retrospective v5.0 proteome dataset.
}
\keyword{datasets}
