% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm_setup_helpers.R
\name{build_formula_string}
\alias{build_formula_string}
\title{Build a formula string from user selections}
\usage{
build_formula_string(
  variables,
  include_intercept = TRUE,
  interactions = list()
)
}
\arguments{
\item{variables}{Character vector of selected variable names}

\item{include_intercept}{Logical, whether to include intercept}

\item{interactions}{List of character vectors, each a pair of variable names}
}
\value{
A formula string (e.g., "~ Treatment + Age + Treatment:Age")
}
\description{
Build a formula string from user selections
}
