% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm_setup_helpers_contrasts.R
\name{build_simple_expr}
\alias{build_simple_expr}
\title{Build a simple contrast expression string (num - den) for limma::makeContrasts.}
\usage{
build_simple_expr(num, den)
}
\arguments{
\item{num}{Character scalar, numerator design-column name.}

\item{den}{Character scalar, denominator design-column name.}
}
\value{
Character scalar, e.g. "treatmentDrug - treatmentVehicle". Empty if
either input is empty.
}
\description{
Build a simple contrast expression string (num - den) for limma::makeContrasts.
}
