% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{contrast_labels_to_list}
\alias{contrast_labels_to_list}
\title{Convert contrast labels to list format expected by stat.testing}
\usage{
contrast_labels_to_list(contrast_labels)
}
\arguments{
\item{contrast_labels}{Character vector of contrasts in "Group1 / Group2" format}
}
\value{
List of character vectors, each with c(numerator, denominator)
}
\description{
Convert contrast labels to list format expected by stat.testing
}
