% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{generate_all_pairwise}
\alias{generate_all_pairwise}
\title{Generate all pairwise contrasts from groups}
\usage{
generate_all_pairwise(groups, bidirectional = TRUE)
}
\arguments{
\item{groups}{Character vector of group names}

\item{bidirectional}{Logical, if TRUE includes both A/B and B/A (default TRUE)}
}
\value{
Character vector of contrasts in "Group1 / Group2" format
}
\description{
Generate all pairwise contrasts from groups
}
