% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{generate_all_vs_multiple_references}
\alias{generate_all_vs_multiple_references}
\title{Generate contrasts for all groups vs multiple reference groups}
\usage{
generate_all_vs_multiple_references(
  groups,
  reference_groups,
  bidirectional = FALSE
)
}
\arguments{
\item{groups}{Character vector of group names}

\item{reference_groups}{Character vector of reference group names}

\item{bidirectional}{Logical, if TRUE includes both directions (default FALSE)}
}
\value{
Character vector of contrasts in "Group / Reference" format
}
\description{
Generate contrasts for all groups vs multiple reference groups
}
\details{
Generates contrasts comparing each non-reference group to each
reference group. Useful when multiple control groups are specified.
}
