% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{generate_all_vs_reference}
\alias{generate_all_vs_reference}
\title{Generate contrasts for all groups vs a reference group}
\usage{
generate_all_vs_reference(groups, reference, bidirectional = FALSE)
}
\arguments{
\item{groups}{Character vector of group names}

\item{reference}{String name of reference group}

\item{bidirectional}{Logical, if TRUE includes both directions (default FALSE)}
}
\value{
Character vector of contrasts in "Group / Reference" format
}
\description{
Generate contrasts for all groups vs a reference group
}
