% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{is_valid_contrast}
\alias{is_valid_contrast}
\title{Check if a contrast is valid for given groups}
\usage{
is_valid_contrast(contrast_label, groups)
}
\arguments{
\item{contrast_label}{String in "Group1 / Group2" format}

\item{groups}{Character vector of available group names}
}
\value{
Logical, TRUE if both groups in contrast exist in groups vector
}
\description{
Check if a contrast is valid for given groups
}
