% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{parse_contrast_label}
\alias{parse_contrast_label}
\title{Parse contrast label to get numerator and denominator groups}
\usage{
parse_contrast_label(contrast_label)
}
\arguments{
\item{contrast_label}{String in "Group1 / Group2" format}
}
\value{
List with elements 'numerator' and 'denominator'
}
\description{
Parse contrast label to get numerator and denominator groups
}
