% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{render_control_group_selector}
\alias{render_control_group_selector}
\title{Render control group selector UI}
\usage{
render_control_group_selector(groups, selected_controls, ns)
}
\arguments{
\item{groups}{Character vector of group names}

\item{selected_controls}{Character vector of currently selected control groups}

\item{ns}{Namespace function from Shiny module}
}
\value{
Shiny HTML tags for the control group selector
}
\description{
Render control group selector UI
}
\details{
Creates a badge/button interface for selecting one or more control
groups. Used with "All vs Control" quick select button.
}
