% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_stat_setup_contrast_helpers.R
\name{render_group_selection_matrix}
\alias{render_group_selection_matrix}
\title{Render group selection matrix UI for one-sample t-test and F-test}
\usage{
render_group_selection_matrix(groups, selected_groups, ns)
}
\arguments{
\item{groups}{Character vector of group names}

\item{selected_groups}{Character vector of currently selected groups}

\item{ns}{Namespace function from Shiny module}
}
\value{
Shiny HTML tags for the group selection interface
}
\description{
Render group selection matrix UI for one-sample t-test and F-test
}
\details{
Creates a checkbox interface where users can select groups to include in the test.
}
