% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_setup_helpers_spectronaut.R
\name{spectronautSetupUI}
\alias{spectronautSetupUI}
\title{UI for Spectronaut preprocessing options}
\usage{
spectronautSetupUI(ns, data_columns)
}
\arguments{
\item{ns}{namespace function from session$ns}

\item{data_columns}{character vector of column names from the uploaded data}
}
\value{
tagList
}
\description{
UI for Spectronaut preprocessing options
}
