% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_setup_helpers_csv-excel-processing.R
\name{validateUniqueIdentifiers}
\alias{validateUniqueIdentifiers}
\title{Validate that identifier column has unique values}
\usage{
validateUniqueIdentifiers(data, identifier_column)
}
\arguments{
\item{data}{data.frame containing the identifier column}

\item{identifier_column}{character name of the identifier column to validate}
}
\value{
\code{TRUE} invisibly if valid; otherwise throws an error
}
\description{
Validate that identifier column has unique values
}
\details{
Errors if the identifier column contains any NA values (all feature IDs
must be non-missing). Also errors on duplicate or empty string values.
}
