% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_lm_results_helpers.R
\name{write_lm_stat_gct}
\alias{write_lm_stat_gct}
\title{Write an LM-results GCT for SSGSEA consumption.}
\usage{
write_lm_stat_gct(df, dir_name, ome)
}
\arguments{
\item{df}{LM results frame.}

\item{dir_name}{Output directory.}

\item{ome}{Ome name; embedded in the filename.}
}
\value{
Invisible path on success, NULL on no-op.
}
\description{
Write an LM-results GCT for SSGSEA consumption.
}
