% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_setup_helpers_spectronaut.R
\name{extract_protigy_id}
\alias{extract_protigy_id}
\title{Extract protigy_id from a semicolon-delimited protein groups column}
\usage{
extract_protigy_id(data, source_column, separator = ";")
}
\arguments{
\item{data}{data.frame}

\item{source_column}{character name of column to split}

\item{separator}{character separator (default ";")}
}
\value{
data.frame with protigy_id column at position 1
}
\description{
Splits each value in source_column by separator, takes the first non-empty
token after trimws(), stores result as protigy_id column at position 1.
}
