% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protigy-package.R
\docType{package}
\name{Protigy-package}
\alias{Protigy}
\alias{Protigy-package}
\title{Protigy: Proteomics Toolset for Integrative Data Analysis}
\description{
A tool for multiomics data analysis and visualization.
}
\author{
\strong{Maintainer}: Natalie M Clark \email{nclark@broadinstitute.org}

Authors:
\itemize{
  \item Stephanie Vartany \email{svartany@broadinstitute.org}
  \item D R Mani \email{manidr@broadinstitute.org}
  \item C Williams \email{wcorinne@broadinstitute.org}
}

}
\keyword{internal}
