% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protigyRevamp-package.R
\docType{package}
\name{protigyRevamp-package}
\alias{protigyRevamp}
\alias{protigyRevamp-package}
\title{protigyRevamp: Revamp of the original Protigy app}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Stephanie svartany@broadinstitute.org Vartany
}

}
\keyword{internal}
