% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protigyRevamp-package.R
\docType{package}
\name{protigyRevamp-package}
\alias{protigyRevamp}
\alias{protigyRevamp-package}
\title{protigyRevamp: Revamp of the original Protigy app}
\description{
A tool for multiomics data analysis and visualization.
}
\author{
\strong{Maintainer}: Natalie M Clark \email{nclark@broadinstitute.org}

Authors:
\itemize{
  \item Stephanie Vartany \email{svartany@broadinstitute.org}
  \item D R Mani \email{manidr@broadinstitute.org}
  \item C Williams \email{wcorinne@broadinstitute.org}
}

}
\keyword{internal}
