% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_setup_helpers_spectronaut.R
\name{split_gene_symbol_column}
\alias{split_gene_symbol_column}
\title{Split a gene symbol column and use the first non-empty value}
\usage{
split_gene_symbol_column(rdesc, column, separator = ";")
}
\arguments{
\item{rdesc}{data.frame (row descriptor)}

\item{column}{character name of column to split}

\item{separator}{character separator (default ";")}
}
\value{
modified rdesc
}
\description{
For each value in rdesc[\link[shiny:column]{shiny::column}], splits by separator and takes the first
non-empty token. Assigns result to rdesc$geneSymbol.
}
